Pass list of unsorted books down to shelf changer to get current shelf
This is more reliable than passing the shelfForBook method down
because the shelfForBook method utilized the filtered books. It was
not guaranteed that the books would be done filtering by the time
the list, book, and shelf changer components were rendered. Assigning
the unfiltered books is faster and almost guaranteed to be done by
the time the child components are rendered.